TMS-231: Add decorator to cache data on a centerpage.
The results of the decorated function will be stored in a dictionary on
the centerpage instance and invalidated at the transaction boundary. The
dictionary keys are built using the optional `keyfunc`, which is called
with `self` as a single argument. By default the instances `__name__`
attribute is used, which works nicely for our (automatic) areas.